services/users/requirements.txt
# bug in werkzeug for docker, use below version
werkzeug==0.14.1
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.7
Flask-Testing==0.7.1
coverage==4.5.3
flask-debugtoolbar==0.10.1